# openapi.actions.yaml
openapi: 3.1.0
info:
  title: Zotero Bridge API
  description: Zotero I/O and workflow layer for ChatGPT/Codex.
  version: 2.0.0
paths:
...
  /v1/papers/add-by-doi:
    post:
      tags:
      - Papers
      summary: Add By Doi
      operationId: addByDOI
      parameters:
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddByDOIRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddByDOIResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/papers/import-metadata:
    post:
      tags:
      - Papers
      summary: Import Metadata
      operationId: importPaperMetadata
      parameters:
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportMetadataRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportMetadataResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/papers/import-discovery-hit:
    post:
      tags:
      - Papers
      summary: Import Discovery Hit
      operationId: importDiscoveryHit
      parameters:
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportDiscoveryHitRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportDiscoveryHitResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/papers/upload-pdf-action:
    post:
      tags:
      - Papers
      summary: Upload Pdf Action
      operationId: uploadPdfAction
      parameters:
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadPdfActionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadPdfResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/items/{itemKey}/attachments:
    get:
      tags:
      - Items
      summary: List Item Attachments
      operationId: listItemAttachments
      parameters:
      - name: itemKey
        in: path
        required: true
        schema:
          type: string
          title: Itemkey
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/attachments/{attachmentKey}:
    get:
      tags:
      - Attachments
      summary: Get Attachment Detail
      operationId: getAttachmentDetail
      parameters:
      - name: attachmentKey
        in: path
        required: true
        schema:
          type: string
          title: Attachmentkey
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/attachments/{attachmentKey}/handoff:
    post:
      tags:
      - Attachments
      summary: Create Attachment Handoff
      operationId: createAttachmentHandoff
      parameters:
      - name: attachmentKey
        in: path
        required: true
        schema:
          type: string
          title: Attachmentkey
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachmentHandoffRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentHandoffResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/attachments/download/{token}:
    get:
      tags:
      - Attachments
      summary: Download Attachment By Token
      operationId: downloadAttachmentByToken
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Attachment download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/library/stats:
    get:
      tags:
      - Library
      summary: Get Library Stats
      operationId: getLibraryStats
      parameters:
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryStatsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/collections:
    get:
      tags:
      - Library
      summary: List Collections
      operationId: listCollections
      parameters:
      - name: start
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Start
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: topLevelOnly
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Toplevelonly
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/tags:
    get:
      tags:
      - Library
      summary: List Tags
      operationId: listTags
      parameters:
      - name: start
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Start
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: topLevelOnly
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Toplevelonly
      - name: collectionKey
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Collectionkey
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/notes/{noteKey}:
    get:
      tags:
      - Notes
      summary: Get Note Detail
      operationId: getNoteDetail
      parameters:
      - name: noteKey
        in: path
        required: true
        schema:
          type: string
          title: Notekey
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Notes
      summary: Update Note
      operationId: updateNote
      parameters:
      - name: noteKey
        in: path
        required: true
        schema:
          type: string
          title: Notekey
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteWriteRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteWriteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Notes
      summary: Delete Note
      operationId: deleteNote
      parameters:
      - name: noteKey
        in: path
        required: true
        schema:
          type: string
          title: Notekey
      - name: X-Zotero-API-Key
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Zotero-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteDeleteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    AddByDOIRequest:
      properties:
        doi:
          type: string
          title: Doi
        collectionKey:
          anyOf:
          - type: string
          - type: 'null'
          title: Collectionkey
        tags:
          items:
            type: string
          type: array
          title: Tags
        requestId:
          anyOf:
          - type: string
          - type: 'null'
          title: Requestid
      additionalProperties: false
      type: object
      required:
      - doi
      title: AddByDOIRequest
    AddByDOIResponse:
      properties:
        status:
          $ref: '#/components/schemas/AddByDOIStatus'
        itemKey:
          type: string
          title: Itemkey
        title:
          type: string
          title: Title
        DOI:
          anyOf:
          - type: string
          - type: 'null'
          title: Doi
      additionalProperties: false
      type: object
      required:
      - status
      - itemKey
      - title
      title: AddByDOIResponse
    AddByDOIStatus:
      type: string
      enum:
      - created
      - existing
      title: AddByDOIStatus
...
    AttachmentDetailResponse:
      properties:
        attachment:
          $ref: '#/components/schemas/AttachmentRecord'
      additionalProperties: false
      type: object
      required:
      - attachment
      title: AttachmentDetailResponse
...
    AttachmentHandoffRequest:
      properties:
        mode:
          $ref: '#/components/schemas/AttachmentHandoffMode'
          default: proxy_download
        expiresInSeconds:
          anyOf:
          - type: integer
            maximum: 86400.0
            minimum: 60.0
          - type: 'null'
          title: Expiresinseconds
        requestId:
          anyOf:
          - type: string
          - type: 'null'
          title: Requestid
      additionalProperties: false
      type: object
      title: AttachmentHandoffRequest
    AttachmentHandoffResponse:
      properties:
        attachmentKey:
          type: string
          title: Attachmentkey
        filename:
          anyOf:
          - type: string
          - type: 'null'
          title: Filename
        contentType:
          type: string
          title: Contenttype
        mode:
          $ref: '#/components/schemas/AttachmentHandoffMode'
        downloadUrl:
          type: string
          title: Downloadurl
        expiresAt:
          type: string
          title: Expiresat
      additionalProperties: false
      type: object
      required:
      - attachmentKey
      - contentType
      - mode
      - downloadUrl
      - expiresAt
      title: AttachmentHandoffResponse
    AttachmentListResponse:
      properties:
        itemKey:
          type: string
          title: Itemkey
        attachments:
          items:
            $ref: '#/components/schemas/AttachmentRecord'
          type: array
          title: Attachments
        count:
          type: integer
          title: Count
      additionalProperties: false
      type: object
      required:
      - itemKey
      - attachments
      - count
      title: AttachmentListResponse
...
    CollectionListResponse:
      properties:
        collections:
          items:
            $ref: '#/components/schemas/CollectionSummary'
          type: array
          title: Collections
        count:
          type: integer
          title: Count
        total:
          type: integer
          title: Total
        start:
          type: integer
          title: Start
        limit:
          type: integer
          title: Limit
        nextStart:
          anyOf:
          - type: integer
          - type: 'null'
          title: Nextstart
      additionalProperties: false
      type: object
      required:
      - collections
      - count
      - total
      - start
      - limit
      title: CollectionListResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ImportDiscoveryHitRequest:
      properties:
        openAlexId:
          anyOf:
          - type: string
          - type: 'null'
          title: Openalexid
        title:
          type: string
          title: Title
        doi:
          anyOf:
          - type: string
          - type: 'null'
          title: Doi
        publicationYear:
          anyOf:
          - type: integer
          - type: 'null'
          title: Publicationyear
        publicationDate:
          anyOf:
          - type: string
          - type: 'null'
          title: Publicationdate
        workType:
          anyOf:
          - type: string
          - type: 'null'
          title: Worktype
        venue:
          anyOf:
          - type: string
          - type: 'null'
          title: Venue
        landingPageUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Landingpageurl
        pdfUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Pdfurl
        isOpenAccess:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Isopenaccess
        abstract:
          anyOf:
          - type: string
          - type: 'null'
          title: Abstract
        authors:
          items:
            $ref: '#/components/schemas/DiscoveryAuthor'
          type: array
          title: Authors
        topics:
          items:
            $ref: '#/components/schemas/DiscoveryTopic'
          type: array
          title: Topics
        attachPdfFromOpenAccessUrl:
          type: boolean
          title: Attachpdffromopenaccessurl
          default: false
        collectionKey:
          anyOf:
          - type: string
          - type: 'null'
          title: Collectionkey
        tags:
          items:
            type: string
          type: array
          title: Tags
        updateIfExists:
          type: boolean
          title: Updateifexists
          default: false
        requestId:
          anyOf:
          - type: string
          - type: 'null'
          title: Requestid
      additionalProperties: false
      type: object
      required:
      - title
      title: ImportDiscoveryHitRequest
    ImportDiscoveryHitResponse:
      properties:
        status:
          $ref: '#/components/schemas/ImportStatus'
        itemKey:
          type: string
          title: Itemkey
        title:
          type: string
          title: Title
        dedupeStrategy:
          type: string
          title: Dedupestrategy
        attachment:
          anyOf:
          - $ref: '#/components/schemas/UploadPdfResponse'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - status
      - itemKey
      - title
      - dedupeStrategy
      title: ImportDiscoveryHitResponse
    ImportMetadataRequest:
      properties:
        itemType:
          type: string
          title: Itemtype
        title:
          type: string
          title: Title
        creators:
          items:
            $ref: '#/components/schemas/CreatorInput'
          type: array
          title: Creators
        abstractNote:
          anyOf:
          - type: string
          - type: 'null'
          title: Abstractnote
        publicationTitle:
          anyOf:
          - type: string
          - type: 'null'
          title: Publicationtitle
        date:
          anyOf:
          - type: string
          - type: 'null'
          title: Date
        doi:
          anyOf:
          - type: string
          - type: 'null'
          title: Doi
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        tags:
          items:
            type: string
          type: array
          title: Tags
        collectionKey:
          anyOf:
          - type: string
          - type: 'null'
          title: Collectionkey
        extra:
          anyOf:
          - type: string
          - type: 'null'
          title: Extra
        updateIfExists:
          type: boolean
          title: Updateifexists
          default: false
        requestId:
          anyOf:
          - type: string
          - type: 'null'
          title: Requestid
      additionalProperties: false
      type: object
      required:
      - itemType
      - title
      title: ImportMetadataRequest
    ImportMetadataResponse:
      properties:
        status:
          $ref: '#/components/schemas/ImportStatus'
        itemKey:
          type: string
          title: Itemkey
        title:
          type: string
          title: Title
        dedupeStrategy:
          type: string
          title: Dedupestrategy
      additionalProperties: false
      type: object
      required:
      - status
      - itemKey
      - title
      - dedupeStrategy
      title: ImportMetadataResponse
...
    LibraryStatsResponse:
      properties:
        totalItems:
          type: integer
          title: Totalitems
        itemTypeCounts:
          items:
            $ref: '#/components/schemas/ItemTypeCount'
          type: array
          title: Itemtypecounts
        collectionCount:
          type: integer
          title: Collectioncount
        tagCount:
          type: integer
          title: Tagcount
        duplicateGroups:
          $ref: '#/components/schemas/DuplicateStats'
        lastModifiedVersion:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lastmodifiedversion
        searchIndex:
          $ref: '#/components/schemas/SearchIndexStats'
      additionalProperties: false
      type: object
      required:
      - totalItems
      - itemTypeCounts
      - collectionCount
      - tagCount
      - duplicateGroups
      - searchIndex
      title: LibraryStatsResponse
...
    TagListResponse:
      properties:
        tags:
          items:
            $ref: '#/components/schemas/TagSummary'
          type: array
          title: Tags
        count:
          type: integer
          title: Count
        total:
          type: integer
          title: Total
        start:
          type: integer
          title: Start
        limit:
          type: integer
          title: Limit
        nextStart:
          anyOf:
          - type: integer
          - type: 'null'
          title: Nextstart
      additionalProperties: false
      type: object
      required:
      - tags
      - count
      - total
      - start
      - limit
      title: TagListResponse
...
    UploadPdfActionRequest:
      properties:
        itemKey:
          anyOf:
          - type: string
          - type: 'null'
          title: Itemkey
        doi:
          anyOf:
          - type: string
          - type: 'null'
          title: Doi
        fileUrl:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Fileurl
        openaiFileIdRefs:
          items:
            $ref: '#/components/schemas/OpenAIFileRef'
          type: array
          title: Openaifileidrefs
        collectionKey:
          anyOf:
          - type: string
          - type: 'null'
          title: Collectionkey
        tags:
          items:
            type: string
          type: array
          title: Tags
        createTopLevelAttachmentIfNeeded:
          type: boolean
          title: Createtoplevelattachmentifneeded
          default: false
        requestId:
          anyOf:
          - type: string
          - type: 'null'
          title: Requestid
      additionalProperties: false
      type: object
      title: UploadPdfActionRequest
    UploadPdfResponse:
      properties:
        status:
          $ref: '#/components/schemas/UploadPdfStatus'
        itemKey:
          anyOf:
          - type: string
          - type: 'null'
          title: Itemkey
        attachmentKey:
          type: string
          title: Attachmentkey
        filename:
          anyOf:
          - type: string
          - type: 'null'
          title: Filename
        contentType:
          type: string
          title: Contenttype
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      additionalProperties: false
      type: object
      required:
      - status
      - attachmentKey
      - contentType
      title: UploadPdfResponse
...